стены, изменение бага со стрельбой,  сохранения,  логика стрельбы для гранатомета, чистка кода

+++ Assets/Scripts/Guns/GunBase.cs
@@ -1,11 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public abstract class GunBase : MonoBehaviour
 {
     [SerializeField] private int _damage;
-    [SerializeField] private int _shootingSpeed;
+    [SerializeField] private float _shootingSpeed;
 
     public float ShootingSpeed { get { return _shootingSpeed; } }
     public int Damage { get { return _damage; } }
